Add button picker and preview
